+++ ae/PlayerData.h
@@ -9,12 +9,12 @@
 
 class PlayerData {
     public:
-        uint_fast8_t Jump, Grounded;
+        uint_least8_t Jump, Grounded;
         VECTOR Last;
         MMD_t MMD;
         PlayerData();
         void Load();
-        void Update(Sphere_t *sObj, uint_fast8_t Destroyed, const int MAX);
+        void Update(Sphere_t *sObj, int_fast16_t Destroyed, const int_fast16_t MAX);
         void Draw(Sphere_t *sObj);
     private:
         float anchor_y;
